refactor(rocr): Inline note bounds helpers into amd_hsa_code.cpp
