+++ projects/rocr-runtime/runtime/hsa-runtime/libamdhsacode/amd_hsa_code.cpp
@@ -46,7 +46,6 @@
 #include <iomanip>
 #include <algorithm>
 #include "core/inc/amd_hsa_code.hpp"
-#include "amd_hsa_note_bounds.hpp"
 #include "amd_hsa_code_util.hpp"
 #include <libelf.h>
 #include "inc/amd_hsa_elf.h"
@@ -67,6 +66,23 @@ namespace rocr {
 namespace amd {
 namespace hsa {
 namespace code {
+
+namespace detail {
+
+// Returns false when claimed_size would cause GetNoteString to read past desc_size.
+inline bool IsNoteStringSizeWithinDescriptor(uint32_t desc_size, size_t field_offset,
+                                             uint16_t claimed_size) {
+  if (desc_size < field_offset) { return false; }
+  return claimed_size <= desc_size - field_offset;
+}
+
+// Bounds a trailing variable-length field within room already carved out of the descriptor.
+inline bool IsNoteStringSizeWithinRoom(size_t room, size_t offset, uint16_t claimed_size) {
+  if (offset > room) { return false; }
+  return claimed_size <= room - offset;
+}
+
+}  // namespace detail
 
     using amd::elf::GetNoteString;
